refactor seeders
