+++ app/seeds/__init__.py
@@ -1,1 +1,2 @@
-from .seeder import IngredientSeeder, BeverageSeeder
+from app.seeds.ingredient_seeder import IngredientTemplate
+from app.seeds.beverage_seeder import BeverageTemplate
